% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{exit}
\alias{exit}
\title{Shortcut to quit function}
\usage{
exit(save = "yes", ...)
}
\arguments{
\item{save}{Default is \code{"yes"}}

\item{...}{Other arguments passed to \code{quit} function.}
}
\description{
Shortcut to quit function
}
\seealso{
\link{quit}
}
