% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{h}
\alias{h}
\title{Shortcut to head function}
\usage{
h(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{arguments to be passed to or from other methods}
}
\description{
Shortcut to head function
}
\seealso{
\link{head}
}
