% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{icat}
\alias{icat}
\title{cat plot on terminal
This function requires iterm terminal and shell integration features.}
\usage{
icat(x, ...)
}
\arguments{
\item{x}{plot object to be printed to terminal.}

\item{...}{Arguments passed to \link{png} function.}
}
\description{
cat plot on terminal
This function requires iterm terminal and shell integration features.
}
