% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{mb}
\alias{mb}
\title{Shortcut to microbenchmark function}
\usage{
mb(...)
}
\arguments{
\item{...}{Arguments passed to \code{microbenchmark} function.}
}
\description{
Shortcut to microbenchmark function
}
\seealso{
\link[microbenchmark]{microbenchmark}
}
