% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{q}
\alias{q}
\title{Shortcut to quit function}
\usage{
q(save = "no", ...)
}
\arguments{
\item{save}{Default is \code{"no"}}

\item{...}{Other arguments passed to \link{quit} function}
}
\description{
Shortcut to quit function
}
\seealso{
\link{quit}
}
