% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{s}
\alias{s}
\title{Shortcut to summary function}
\usage{
s(object, ...)
}
\arguments{
\item{object}{an object for which a summarhy is desired.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Shortcut to summary function
}
\seealso{
\link{summary}
}
